+++ src/parent_sync.rs
@@ -6,7 +6,7 @@ use bevy::{
     prelude::*,
 };
 
-use crate::{has_authority, AppReplicationExt};
+use crate::{client::ClientSet, has_authority, prelude::ServerSet, AppReplicationExt};
 
 pub struct ParentSyncPlugin;
 
@@ -17,13 +17,12 @@ impl Plugin for ParentSyncPlugin {
     fn build(&self, app: &mut App) {
         app.register_type::<Option<Entity>>()
             .replicate::<ParentSync>()
+            .add_systems(PreUpdate, Self::sync_system.after(ClientSet::Receive))
             .add_systems(
                 PostUpdate,
-                (
-                    (Self::update_system, Self::removal_system).run_if(has_authority()),
-                    Self::sync_system,
-                )
-                    .chain(),
+                (Self::update_system, Self::removal_system)
+                    .run_if(has_authority())
+                    .before(ServerSet::Send),
             );
     }
 }
@@ -68,6 +67,8 @@ impl ParentSyncPlugin {
 /// Updates entity parent on change.
 ///
 /// Removes the parent if `None`.
+/// The component captures changes in `PostUpdate` on server before sending
+/// and applies them on `PreUpdate` after receive on clients or scene deserialization.
 #[derive(Component, Default, Reflect, Clone, Copy)]
 #[reflect(Component, MapEntities)]
 pub struct ParentSync(Option<Entity>);
@@ -92,8 +93,13 @@ mod tests {
         let mut app = App::new();
         app.add_plugins((ReplicationCorePlugin, ParentSyncPlugin));
 
-        let child_entity = app.world.spawn(ParentSync::default()).id();
+        let child_entity = app.world.spawn_empty().id();
         app.world.spawn_empty().add_child(child_entity);
+
+        app.add_systems(Update, move |mut commands: Commands| {
+            // Should be inserted in `Update` to avoid sync in `PreUpdate`.
+            commands.entity(child_entity).insert(ParentSync::default());
+        });
 
         app.update();
 
@@ -111,10 +117,17 @@ mod tests {
         let parent_entity = app.world.spawn_empty().id();
         let child_entity = app
             .world
-            .spawn(ParentSync(Some(parent_entity)))
+            .spawn_empty()
             .set_parent(parent_entity)
             .remove_parent()
             .id();
+
+        app.add_systems(Update, move |mut commands: Commands| {
+            // Should be inserted in `Update` to avoid sync in `PreUpdate`.
+            commands
+                .entity(child_entity)
+                .insert(ParentSync(Some(parent_entity)));
+        });
 
         app.update();
 
@@ -123,7 +136,7 @@ mod tests {
     }
 
     #[test]
-    fn hierarchy_set() {
+    fn update_sync() {
         let mut app = App::new();
         app.add_plugins((ReplicationCorePlugin, ParentSyncPlugin));
 
@@ -139,7 +152,7 @@ mod tests {
     }
 
     #[test]
-    fn hierarchy_unset() {
+    fn removal_sync() {
         let mut app = App::new();
         app.add_plugins((ReplicationCorePlugin, ParentSyncPlugin));
 
@@ -160,7 +173,8 @@ mod tests {
     }
 
     #[test]
-    fn scene_hierarchy_set() {
+    #[ignore = "will work if Bevy move scene spawning to `PreUpdate`"]
+    fn scene_update_sync() {
         let mut app = App::new();
         app.add_plugins((
             AssetPlugin::default(),